requirements.txt
discord.py >= 1.7.3
aiohttp >= 3.7.4
cairosvg >= 2.5.2
chess >= 1.6.1
matplotlib >= 3.4.2
motor >= 2.4.0
youtube_dl >= 2021.6.6
pynacl >= 1.4.0
openai >= 0.10.1